fix: update production quantity type to int and add validation contraints

+++ src/main/java/afsj/efm/production/application/dtos/ProductionResponse.java
@@ -3,11 +3,11 @@ package afsj.efm.production.application.dtos;
 import java.time.LocalDate;
 
 public record ProductionResponse(
-         Long id,
-         Long areaId,
-         Long productId,
-         Long quantity,
-         String observation,
-         LocalDate createdAt
+        Long id,
+        Long areaId,
+        Long productId,
+        int quantity,
+        String observation,
+        LocalDate createdAt
 ) {
 }
